#Jira BRS-20 - Add pseudo folders to Xena - automatically create folders from configuration for new orders + refactor
